[SW-1539][Hotfix] Avoid boiler-plate code when introducing new test suite in PySparkling
