+++ py/tests/integration_yarn/test_integ_yarn.py
@@ -14,18 +14,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import pytest
-from pyspark.sql import SparkSession
 
-from tests import unit_test_utils
+"""
+Integration tests for pySparkling for Spark running in YARN mode
+"""
+
+from tests.integ_test_utils import *
 
 
-@pytest.fixture(scope="module")
-def spark(spark_conf):
-    conf = unit_test_utils.get_default_spark_conf(spark_conf)
-    return SparkSession.builder.config(conf=conf).getOrCreate()
+def test_xgboost_medium(integ_spark_conf):
+    return_code = launch(integ_spark_conf, "examples/scripts/tests/xgboost_test_medium.py")
+    assert return_code == 0, "Process ended in a wrong way. It ended with return code " + str(return_code)
 
-@pytest.fixture(scope="module")
-def prostateDataset(spark):
-    return spark.read.csv("file://" + unit_test_utils.locate("smalldata/prostate/prostate.csv"),
-                          header=True, inferSchema=True)
+
+def test_chicago_crime(integ_spark_conf):
+    return_code = launch(integ_spark_conf, "examples/scripts/ChicagoCrimeDemo.py")
+    assert return_code == 0, "Process ended in a wrong way. It ended with return code " + str(return_code)
